Style Main component
Created a grid layout, with the leftmost area
reserved for the sidebar and the relevant channels,
links, and buttons, and the remaining portion dedicated
to the chatroom.

Styled user messages (using flex), added mouse
hovering effect over messages, and sized and spaced
as close to the actual Slack as I could.

Customized the scrollbar to the Attack color scheme,
rounded the edges, and added conditional rendering.

Styled the message form, with spacing, coloring, sizing,
and ::focus as close to the actual Slack as I could.