Restyle splash page
The splash page was given a significant makeover, both to account
for Slack's very own redesign of its brand and splash page, and
because the previously written code was very messy, inefficient,
and burdensome.

The new layout consists simply of two components: the navbar, and
the main splash body. The navbar is a flexed navbar with a font
more closely resembling that on Slack's frontpage. The body is a
grid layout that exactly mirrors that of Slack's, minus the fancy
image slideshow.