Style navbar, sidebar, and message items
Main navbar was styled to include icon to
display channel headcount and the actual number of
users in a channel, as well as the channel description.

Message items were restyled to accomodate space for a user
profile photo.

Channel items on the sidebar were given extra padding
to more closely mirror Slack's.

Message form was vertically centered in its grid cell.